move viewer dialog to a separate folder
