+++ ng2-components/ng2-alfresco-viewer/src/components/viewer-dialog/viewer-dialog.component.scss
